chore(Tests): Move tests to be next to the file they test
