+++ src/guildHandler/Components/Data/Settings/config/permissionConfig.ts
@@ -2,16 +2,16 @@
 export type PermissionsConfig = {
 	[key: string]: Array<string>;
 };
-export const PERMISSIONS_DEFAULT = {
+export const PERMISSIONS_DEFAULT = Object.freeze({
 	everyone: [
-		'join', 
-		'play', 
-		'pause', 
-		'resume', 
+		'join',
+		'play',
+		'pause',
+		'resume',
 		'stop', 'leave',
 		'skip', 'next',
 		'repeat', 'repeat-song', 'rs',
-		'repeat-queue', 'rq', 
+		'repeat-queue', 'rq',
 		'shuffle', 'toggle-shuffle',
 		'show-queue', 'sq',
 		'clear-queue', 'cq',
@@ -21,4 +21,4 @@ export const PERMISSIONS_DEFAULT = {
 		'autoplay', 'toggle-autoplay'
 	],
 	admin: ['set-channel']
-};
+});
