--- src/guildHandler/Components/tests/AudioProcessor.test.ts
@@ -1,23 +1,0 @@
-import { should } from 'chai';
-should();
-import sinon from 'sinon';
-
-before(() => {
-	//
-});
-
-beforeEach(async () => {
-	//
-});
-
-afterEach(() => {
-	//
-});
-
-after(() => {
-	//
-});
-
-describe('', () => {
-	//
-});
